requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.28
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.3
greenlet==3.0.1 